Time: 0 ms (100%), Space: 8.4 MB (99.2%) - LeetHub

+++ 0069-sqrtx/0069-sqrtx.cpp
@@ -5,7 +5,7 @@ public:
         int e = x;
         while(s<=e){
             int mid = e-(e-s)/2;
-            int sq = mid *mid;
+            long long sq = (long long)mid *mid;
             if(  sq== x){
                 return mid;
             }
